backend/requirements.txt
# Web framework
FastAPI==0.110.0
uvicorn[standard]==0.29.0

# Validation
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

# Database
SQLAlchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP client
httpx==0.27.0

# Cache / queue
redis==5.0.3
celery==5.3.6

# Logging
structlog==24.1.0

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
anyio==4.3.0
